pyproject.toml
[tool.poetry]
name = "concurrent-crawler"
version = "0.1.0"
description = "multithreaded, multiprocess web crawler and scraper"
authors = ["Eric Willenson <eric.willenson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
supabase = "^1.2.0"
requests = "^2.31.0"
html2text = "^2020.1.16"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
